WIM-84 update style and popups
