+++ src/components/Header.jsx
@@ -1,10 +1,11 @@
 import React from 'react'
+import logo from '../assets/0012p000041aQ4LAAU.png'
 
 function Header() {
   return (
-    <div className="header"> <h1>
-      Working Talent Book Rental
-    </h1></div>
+    <div className="header">
+      <img src={logo} alt='Header' className='header-img' />
+    </div>
    
   )
 }
