+++ src/components/Login.jsx
@@ -3,10 +3,16 @@ import { Link } from "react-router-dom";
 
 function Login() {
     return (
-        <div className="login-container">
-            <label>Naam:</label><input type='text' />
-            <label>Wachtwoord:</label><input type='text' />
-            <div className="login-button"><Link to="main">Login</Link></div>
+        <div className="input-container">
+            <div className="login-header">
+                <h4>INLOG SCHERM</h4>
+                <h2>Boeken reserveren</h2>
+            </div>
+            <div className="login-input">
+                <input type='text' placeholder="E-mailadres*" />
+                <input type='text' placeholder="Wachtwoord*" />
+                <div className="login-button"><Link to="main">Login</Link></div>
+            </div>
         </div>
     )
 }
